requirements.txt
openai>=1.40.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
docx2txt>=0.8
PyPDF2>=3.0.0
